# feat(PL-3517): cache rendered values