… go.sum
… internal/github/github_mock.go