### Cache invalidation logic

Changes to:

```
joy.yaml => full cache invalidation
environments/<env>/env.yaml => env cache removed
environments/<env>/releases/<anything>.yaml => item removed
```

### Benchmarks

On my local machine. Note that a `git pull` is involved so that
can add a fair amount of variability/noise.

```
Benchmark After: curl -X POST 127.0.0.1:8080/api/v1/getparams.execute -H 'Authorization: Bearer token' -d {} > /dev/null
Time (mean ± σ):     513.1 ms ±  63.0 ms    [User: 4.6 ms, System: 8.6 ms]
Range (min … max):   438.2 ms … 634.3 ms    10 runs

Benchmark Before: curl -X POST 127.0.0.1:8080/api/v1/getparams.execute -H 'Authorization: Bearer token' -d {} > /dev/null
Time (mean ± σ):      1.609 s ±  1.408 s    [User: 0.004 s, System: 0.006 s]
Range (min … max):    1.078 s …  5.610 s    10 runs
```